+++ packages/libredwg-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libredwg-converter
 
+## 3.1.6
+
+### Patch Changes
+
+- feat: add missed property dimBlockPosition for class AcDbDimension
+- Updated dependencies
+  - @mlightcad/data-model@1.3.6
+
 ## 3.1.5
 
 ### Patch Changes